feature(mine): Threshold value for guessing
With the introduction of the preprocessing step in guessing, it's no
longer free to start guessing as the preprocessing step takes some time
(at least 1 minute on a very fast CPU). This addresses a part of this
problem by allowing the user to set a threshold value under which
guessing will not happen.

Cf. #650.