ci: Skip flaky async tests
In CI, skip tests related to the asynchronous cancelability of
guess-preprocessing. The tests in question trigger an event that
should happen within a given time frame, but there is no good
heuristic for determining the right time frame. It depends on the
clock frequency of the current machine, the ratio of the CPU time
to user time it gets, the number of other processes that are running,
etc. It is better to not pretend that we have a CI failure when the
out-of-time-frame event is due to programmer ignorance of machine
vagaries instead of errors in need of debugging.