feat(API): Build infrastructure for transparent transactions
By default, transactions are private. However, if the user opts in,
they can disclose essential transaction details like the UTXOs that
are consumed and produced along with the commitment randomness that
generates the `RemovalRecord`s and `AdditionRecord`s, respectively.
In this event, the disclosed information suffices for a transparent
audit.

This commit adds a field `transparent` of type `bool` to the API's
`TransactionDetailsBuilder`, along with a self-consumer constructor
helper to set its value. If the boolean is set, then the builder will
produce a list of `TransparentInput`s and `UtxoTriple`s, matching the
UTXOs that are consumed and produced by the transaction. The encoding
of these lists is then included as an announcement in the transaction.

Struct `TransparentInput`s is new; no existing struct offered the
required fields without superfluous data. In contrast, struct
`UtxoTriple` matches exactly with what `TransparentOutput` would have
been, so rather than duplicating it, this commit extends its
functionality so that it can serve multiple needs at once.