refactor(mine_loop): Fix block timestamp
Ensure block timestamp is never earlier than transaction timestamp. This
is achieved by setting the block timestamp to the transaction timestamp.
Prior to this commit, the block transaction timestamp was set to that of
the coinbase transaction which is wrong because the merger of a new
transaction might bumb the transaction timestamp.

Co-authored-by: Alan Szepieniec <alan@neptune.cash>