refactor!(peer_loop): Use shared const for future dating limit
The peer loop rejects transactions that are too far into the future.
Previously, it used its own const for this check but now it uses a const
that's defined in the block module such that the same const used in the
consensus mechanism is used in the peer loop. That will protect
composers from receving into their mempool transactions which will
create invalid block proposals.

Also: Change a `>` into a `>=` in peer loop to make the transaction
timestamp check the same as the consensus rule check in Bloc::is_valid.

Co-authored-by: Alan Szepieniec <alan@neptune.cash>