feat(guessing): Only switch proposals when reward delta meets threshold
Add a CLI argument that determines when a guesser switches from one
proposal to another. Since PoW guessing requires an expensive
preprocessing step before actual guessing can begin, a guesser should
not change proposals unless the new proposal has a sufficiently large
improvement in the reward.

The default value of this parameter is set to 17 %.

This closes #648.