chore: Upgrade dependencies
Upgrade dependencies, resolve errors, and make the clippy linter happy.
The disallowing of explicit deref rules `explicit_deref_methods` is
deactivated since the `proptest` crate that we rely *heavily* on does
not play ball with it -- so the rule is either deactivated globally, or
we would have to deactivate it many places manually which would clutter
the codebase.

Deletes two unused files, chunk.rs which was moved to the
`removal_record` directory without deleting the file at the old location
a snippet for auditing the end-state of the VM after running a consensus
program, since this snippet was never used.