test(Block): Fix generator for specified block height
Constrain AOCL leaf count to a value below 2^{63} since a helper
function otherwise experiences an overflow.

This problem in the test-case generator was only discovered because we
upgraded the `proptest` library and that has probably
invalidated/changed the deterministism that we used previously. As a
side effect, some new stored proofs need to be generated.