test: Ensure deterministic STARK proofs
Ensure deterministic proofs in tests by setting the RNG seed used by the
Triton VM prover. This breaks zero-knowledge of the proofs used for
tests. This is motivated by the block hash depending on the block proof
meaning that if two different machines start a test simultaneously, they
would otherwise not generate the same proofs and, worse, will disagree on
block hashes and mutator sets if these proofs are used for blocks.

This closes #657.