feat(mine_loop): Update tx if no synced in mempool
If the composer finds no synced transaction in the mempool, they now
look in the mempool for a non-synced single proof-backed transaction
that can then first be updated and then merged.

A transaction included in a block must have its merge bit set which means it must go through at least one merge.

Previously two paths to a block transaction existed:
a) Get a transaction from mempool and merge it with the coinbase transaction
b) If no synced transaction exists in the mempool, make a nop transaction and merge it with the coinbase transaction.

The PR introduces a third path to a block transaction:
c) If an unsynced transaction exists in the mempool, update the transaction and merge it with the coinbase transaction.

Since updating is usually a lot faster than making a new single proof-backed transaction (padded height 2 19 vs 2 21 ) this is beneficial to the composer, as it allows them to compose faster. And this change benefits transaction throughput, so it's a win-win.