feat: Add RPC endpoint for addition records with AOCL indices
Add new RPC endpoint `addition_record_indices_for_block`, which takes a block
digest and returns a hash map containing that block's addition records as keys
and, if the block is canonical, those addition records' AOCL leaf indices as
values. If the block is not canonical, the values are None.

This feature is useful for explorers rendering transparent transactions. The
outputs of the transaction can now have clickable links sending them to the
corresponding UTXO page.