feat: padded height specific tvm env vars
Allow the user to specify a list of environment variables for a given
padded height. This allows the user to build bigger proofs than they
otherwise could by ensuring that very big proofs can be generated
without running out of RAM. On our machines, values like
"24:RAYON_NUM_THREADS=50 TVM_LDE_TRACE=no_cache" allow for the
construction of proofs of a log2 padded height of 24, a proof generation
that would otherwise use more than the available RAM. Powerusers can
play around with these values to find their optimal setup.

Also: Now that we actually have the proof's algebraic execution trace
(AET), we can use it to check if proof complexity is exceeded.
Previously, we had to rely on the processor table to check if complexity
was exceeded but the processor table doesn't necessarily dominate. In
fact, in a recursive setting, the processor table is only ~200k rows per
proof verification whereas the memory table is ~300k rows per
verification. So this new added check will guarantee that the RAM of the
machine is not exhausted during proving, provided that the user knows
the max log2 padded height they can handle.