feat(rpc_server): Add CLI command and endpoint for upgrading specified mempool-tx
Add an endpoint that specifying a transaction kernel ID for a
transaction in the mempool for which a proof upgrade will be performed.
The endpoint automatically chooses the right upgrade to ease the
inclusion of the transaction into a block. Only "update" or "raise" can
be created this way, the command will not spawn "merge" upgrades.