refactor: Factor out inner prove function for testability
Now that the way we invoke Triton VM's prove involves a bigger setup,
the `main` function of the Triton VM-invoking binary is broken up. For
now, only a happy-path vanilla path is tested but this structure allows
for more thorough testing later -- for example with the setting of
environment variables.