refactor: Remove unfinished RPC endpoint
Remove endpoint for tx proof upgrading.

Allowing external programs to upgrade transaction proofs sounds like a
great idea to me -- I think that was what this endpoint intended. But it
wasn't finished and was marked as such. Also: Removing it allows us to
make `Mempool::get_mut` private and test-only which is good since this
function breaks with the rule that all mempool functions should go
through `GlobalState` to allow the state to catch mempool events.