refactor(`OutputFormat`): Drop unused variant
Variant `TxOutput` of `OutputFormat` was not used and is now dropped, along with
an unused `From<TxOutput>` implementation.

The resulting enum captures a smaller set whose elements have more in common.
For instance, all variants have a `ReceivingAddress` (but `TxOutput` did not).
This commonality leads to less special casing and more code reuse.