refactor(`PrimitiveWitness`): Drop `From<&TransactionDetails>` implementation
The `From` conversion was used in one place:

```
-        self.into()
+        PrimitiveWitness::from_transaction_details(self)
```

There is a substantial difference between the types `TransactionDetails` and
`PrimitiveWitness`. The conversion is far from obvious. So the destination type
of `transaction_details.into()` is far from obvious. Making only the dedicated
function available signals to the user that something non-trivial is going on.