refactor(`TransactionOutput`): Move, rename, and extend
Add new directory and module `parser`, which contains structs that exist for
parsing command-line arguments. For now it contains only one struct,
`Beneficiary` which inherits the responsibilities of now-defunct and
hence-dropped `TransactionOutput`. This struct is furthermore extended with an
extra field, `release_date`, which if set generates a time-locked output.

Also: add parsing tests. And test dependencies.