refactor(`OutputFormat`): Drop release date optionality
Variant `OutputFormat::AddressAndAmountAndMaybeReleaseDate` used to have an
`Option<Timestamp>`. Now it is called `AddressAndAmountAndReleaseDate` and has
just a `Timestamp`. The option is already implicit in possibility of choosing
another variant. Phrased differently, `AddressAndAmount(addr, amt)` and
`AddressAndAmountAndMaybeReleaseDate(addr, amt, None)` are the same and we do
not need two different ways of encoding the same thing.