test: Add integration test of time-locked expenditure
The new test verifies that Alice can send funds and attach a time-lock, and Bob
will not be able to spend them before the time-lock expires.

Also: modify API interfaces to pass and use the given timestamp, and avoid using
`Timestamp::now()`. This change is needed for the test to simulate transactions
being initiated one year into the future, but more importantly the API should
support the building of transactions ahead of time.