refactor(mempool): Use MSA tip digest to check for MSA-updated txs
Previously we just used a different value of the mutator set digest to
check if a transaction represented a mutator set update from a
transaction already in the mempool. But since we have the acutal tip's
mutator set hash as a field in the mempool, we can improve this check to
be whether the new transcation has a mutator set hash that matches the
tip's hash.