refactor!(CLI): Allow optional file argument
Commands `send-to-many` and `send-transparent` now have an optional `--file`
argument. If set, the beneficiaries are read from there instead of the required
`outputs` argument, which can now be empty but only if the `--file` is set. If
both a non-empty raw argument and a file are provided, the arguments will be
rejected so as to avoid confusion between which one was chosen.

Mixing required and optional arguments has the unfortunate side-effect of
specifying an ambiguous grammar which clap obviously cannot handle. To fix this
issue, the `--fee` argument must now be spelled out (*i.e.*, by including
`--fee`), whereas previously the last argument was automatically the fee if it
could be parsed as such.