refactor(mempool): Check conflict cache before accepting tx
Rename mempool method that checks if tx is already in mempool to
`accept_transactions`, and make this method also check the conflict
cache. If a transaction is in the conflict cache, it should not be
accepted from a peer.

To achieve this efficiently, the conflict cache now has a hash map with
transaction kernel IDs as keys that can be used for fast lookups.

Harden some tests.