fix!(Tx-API): Specify timestamp
When generating `TransactionDetails` through the API's `generate_tx_details`,
the timestamp would be set to `Timestamp::now()`.

```
        TransactionDetailsBuilder::new()
            .timestamp(Timestamp::now())
            // ...
```

This assignment breaks tests that are sensitive to the timestamp, by setting
field to what might be the wrong value.

A superficial fix is to expand the function signature to include a `Timestamp`
argument, and percolate the interface changes.

However, this commit opts instead to kill `generate_tx_details`, along with its
like-named RPC endpoint.

Motivation: besides the RPC endpoint, it was used in one test and no-where else.
Besides being under-used it was under-documented also:
```
-    /// todo: docs.
```

More importantly, after expanding the argument list with a `Timestamp`, the
inputs to this function would be almost identical to the fields on
`TransactionDetails`. The only relevant difference comes from the change output
UTXO, and that's  only if the input-sum is different from the output-sum and
that's only the case if the change UTXO was not already added to the
`tx_output_list`.

Struct literal syntax is an adequate interface for instantiating a
`TransactionDetails` object from its constituent components. If the API user
must obtain the change output through the API, then a dedicated function
tailor-made for this purpose is preferable.

Hat-tip-to: @sword_smith for helping to track down the cause of the test failure
that led to this resolution.