refactor: Make RegTest launch data constant
Make constant and same as all other networks. This is done to allow for
the reuse of the PoW solution added in the `can_find_valid_pow_solution`
integration test as this test builds a block proposal with height 1 and
then finds a PoW solution for it. If the RegTest genesis block keeps
changing then so does the PoW puzzle for a block of height 1 and this
would invalidate the stored/cached PoW solution, meaning that the cache
would have to be updated whenever the genesis block of the RegTest
network was updated.

This change does not seem to have broken anything or invalidated any
proofs since RegTest uses mocked proofs throughout, so they are not
cached on the proof server.