fix: ensure block data written before db
fixes #471

The issue was that block records were being written to DB before written to
blkXX.dat files on disk, which resulted in mis-match when power is lost during a
call to ArchivalState::store_block().  Upon startup, a SIGBUS occurs when
reading the mmap'd data.

Fixes:

1. root cause: calls mmap.flush() after writing block to ensure it is written to
disk before writing matching record to DB.

2. avoid SIGBUS: verifies we do not attempt to read past end of file.
Returns error with diagnostic info if so, which gets logged.

Co-authored-by: sword-smith <thor@neptune.cash>