refactor(RPC): Make RPC endpoint `set-tip` asynchronous
Setting tip (to stored block) can take a long while if you are sitting
on a large number of UTXOs. If the method takes too long, >10s, the
RPC server will abort the process. But 10s might not be long enough!
To avoid this problem, the tip setting logic is made asynchronous. The
RPC server returns an empty result immediately, and the main loop is
charged with setting the tip to the given digest at its earliest
convenience.