feat: add flag --restrict-peers-to-list
When present the node will avoid connecting to any peer that is not included in
a --peer flag, nor will it accept a connection from any other peer, nor will it
perform peer discovery.

It essentially turns --peer into a whitelist.

One use case is to have a single node on a subnet that connects to the outside
world while all the subnet nodes connect to eachother.  Thus reducing bandwidth
over external link and improving message propagation times to all local nodes.