feat: log guessing rate and estimates
logs guessing information every 30 seconds.  here is a sample:

the "#est#" prefix aids with grepping for multi-line log entries.

"with competition" estimates how long it would take to find a block at present
difficulty with others guessing, so a block is produced every 10 minutes.
It factors in time spent initializing mem for each block.

"in isolation" estimates how long it would take to find the current block if
isolated from the network, so the mem init would only occur once and all the
rest of the time is spent on guessing.