feat: Add tx filter for proof upgrading
To avoid double work if the same entity is controlling multiple nodes
that are performing proof upgrades, a filter is introduced. The filter
looks at the `TransactionKernelId` value and performs a modulus
operation and compares the remainder with that set in the filter. If the
transaction is not a match, no proof upgrading is done (unless the node
has a financial interest in the transaction).

If an entity is controlling three proof upgrading nodes, values for this
filter could be set to "3:0" on the 1st machine, "3:1" on the 2nd, and
"3:2" on the 3rd machine. This way, the entity avoids that the machines
do duplicate work.