chore: Restructure directories related to benchmarks (1/2)
Concretely, relocate `benches/*.rs` to `benchmark/benches/`. Also,
instruct `Cargo.toml` to find the benchmarks in
`benchmark/benches/<file>.rs` using extra lines like the following in
`[[bench]]` sections:

```
  [[bench]]
  name = "sync_atomic"
+ path = "benchmark/benches/sync_atomic.rs"
  harness = false
```

This line will be required from all benchmark programs, including
future ones.

This practice goes against standard practice in rust, which is for
benchmark programs to live in `benches/`. However, this deviation is
warranted by the non-standard practice already in use of committing
produced benchmark reports to the repository. An argument for adhering
to rust standard practice must involve removing the benchmark reports
from the repository (which, to be clear, is *not* the intention of
this commit).

If one deviation from standard practice is allowed, then let us at
least avoid cluttering the root directory with multiple children
directories named and themed along similar lines. Specifically,
directories `benches/` and `benchmarks/` present one more line than
necessary to read across if you are looking for something else; and
if you *are* looking for a benchmark program --or a benchmark report--
then it is not even clear which of these directories you should be
looking in.

The objective of this commit is to collocate related directories
`benches/` and `benchmarks/` into the same well-named parent
directory, `benchmark/`. The end goal is to have the following
directory structure:

 - `benchmark/`
    - `benchmark/benches/` (the benchmark programs `*.rs` live here)
    - `benchmark/reports/` (the tasm benchmark reporst `*.json` live
      here)

However, rather unfortunately, the name of directory `benchmarks/` is
hardcoded in repository `tasm-lib`, file `snippet_bencher.rs`,
function `write_benchmarks`, even for dependencies (like
`neptune-core`) that call it. And so the second part of this commit,
wherein directory `benchmarks/` is renamed to `benchmark/reports/`,
 will have to wait for a future twin commit to this one. That twin
commit depends on a change and version bump in `tasm-lib`.