chore: Move directory `models/state/` out one level
There is a natural semnatic separation between "models", which are
abstract classes of things; and "state" which is a concrete
instantiation of them. This natural semantic separation is now
reflected in the directory structure.

Furthermore, a lopsided fraction of development time so far has been
spent in `models/state/`, suggesting that this directory is important
enough to be promoted to first-class citizen.