chore: Move `config_models/` to `application/config/`
Also, create directory `application/`.

The directory `config_models/`, now `application/config/`, contains
data structures used to configure the application. "Application
configuration" (counted with variations on it) are a more
well-established term to describe these data structures than
"configuration models" (including variants) is. Siding with standard
practice reduces confusion and orientation time.

Moreover, a negligible fraction of development time was spent in
`config_models/`, suggesting that it might not be a good fit for first
class citizenship (*i.e.*, living in the root directory).

I cite the following sources to justify the position that "application
configuration" is more well-established than "configuration models"
(both counted with variants):

 - IBM cloud documentation, "What is App Configuration?" [1]
 - configu.com, "What is Application Configuration?" [2]
 - Microsoft ASP.NET documentation [3] consistently uses the term
   "application configuration" (in prose; in code it prefers
   `appsettings`).

To compound the confusion of the term "configuration models", it is
typically used in abstract computer science to denote, well, abstract
models for representing and manipulating data in the context of
computational tasks. In this context, it is related to "configuration
space" and not to "application". [4]

[1]: https://cloud.ibm.com/docs/app-configuration?topic=app-configuration-ac-about
[2]: https://configu.com/blog/application-configuration-a-practical-guide/
[3]: https://learn.microsoft.com/en-us/aspnet/core/fundamentals/configuration/?view=aspnetcore-9.0
[4]: https://www.sciencedirect.com/topics/computer-science/configuration-model