chore: Create directory `application/control/`
The notion "application control" captures logic related to directing
the application's control flow, whether through event loops or helper
functions.

Besides creating the `control/` directory, this commit moves the
following modules there:

 - main loop
 - peer loop
 - mine loop

While the RPC server also features an event loop, this event loop
exists only for the purpose of responding to RPC calls and contributes
little to orchestrating the application's overall control flow. As
such, the RPC server and related files belong in a different
directory, to be created in a future commit.