chore: Move RPC-related files to `application/rpc/`
Also, given that these files now live in directory `rpc/`, drop the
prefix `rpc_`.

The RPC layer has to do with enabling the user to interface with the
application. It has little to do with the logic of synchronizing state
to the rest of the network, or with participating conformantly with
the blockchain protocol. As such, it is subordinate to the category
"application". This commit reflects that subordination in the
directory structure.