chore: Rename directory `application/control/` to `application/loops/`
New name reflects contents and purpose more concretely: application loops,
running concurrently, responding to events whenever they appear.

The old name is rather abstract. "Control" can mean a lot of things. Moreover,
the motivation for avoiding the word "loop" originally was because the RPC loop
does not live in this folder. Except, the RPC loop does not even live in this
repository: it is part of the RPC library. So that original motivation is void.

Co-authored-by: Thorkil Schmidiger <thor@neptune.cash>