chore: Rename `models/` to `protocol/` and `blockchain/` to `consensus/`
The Protocol is what determines what are and are not conformant interactions for
network participants. This category includes peer-to-peer messages but also, but
also the data types that these messages transfer and pertain to -- including, in
particular, all messages that contribute to the Blockchain.

Consensus deals with validity and canonicity. It contains the blockchain data
structures as well as the tasm programs that define validity. (Conceivably there
may be tasm programs at some point that do not contribute to consensus but for
now there aren't any.)

Co-authored-by: Thorkil Schmidiger <thor@neptune.cash>