chore: Kill type alias for `Tip5`
Use `Tip5::hash` instead of `Hash::hash` and analogously for `hash_varlen`.

Type alias `Hash` provided some genericity which might come in handy in case we
want to switch from Tip5 to another hash function or in case we want to use a
different hash function within one submodule. However, neither event is even
remotely likely and we will deal with the fallout in case it does come to pass.

As an unintended benefit, this commit resolves confusion between `Tip5` and
`std::hash::Hash`.