chore: Move `database.rs` to `state/`
File in question contains logic for storing blocks and therefore relates to
State more than Protocol.

It is tempting to put it into Application. However, the category application
is defined as making abstraction of the particular blockchain protocol and the
particular state. The file in question contains concrete logic unique to this
blockchain protocol's blocks that is unlikely to carry over to a similar
application for a different protocol. Given the concreteness, State is the
better fit.