feat: Add `block-notify` CLI flag option
Add the option of defining a command that is invoked every time the
tip changes. This command is executed whenever the tip changes,
regardless of it being because a new tip was mined locally, it was
received from a peer, or a new tip was set from a stored tip manually.

Intended to closely model Bitcoin Core's `blocknotify` CLI flag.

Unix-test included. We could consider generalizing this test to work for
MacOS and Windows also.