refactor(composer_parameters): Arbitrary coinbase distribution
Prepare data structure `ComposerParameters` to allow for arbitrary
coinbase distributions. The outputs of the coinbase transaction no
longer has to go to one address but can be distributed according to the
new data structure `CoinbaseDistribution` which is a list of
`CoinbaseOutput` where fractions, addresses, and the presence of
timelocks can be specified.

The endpoints/API to set this `CoinbaseDistribution` will be added in
the next commit.