feat(rpc_server): Allow overriding the default coinbase distribution
Add two RPC server endpoints `set_coinbase_distribution` and
`unset_coinbase_distribution` that allow the operator to control the
coinbase distribution of all block proposals originating from their
node.

Adds a field to `MiningState` that holds the value set for the coinbase
distribution. If no value is set, the "default" distribution is made,
the one prior to this commit, if the value is set, then that
distribution is used.