feat(mutator_set): Privacy-preserving membership recovery
Add a method `restore_membership_proof_privacy_preserving` to
`ArchivalMutatorSet` that returns the required data to restore a mutator
set membership proof, without the caller leaking any privacy other than
the fuzzy timestamp that is present when the input gets mined. This
works by taking an absolute index set and returning all the AOCL
authentication paths that this absolute index set could be associated
with.

Adds RPC endpoints for this process also, and a way to extract the
actual membership proof from the returned data.

Most complicated function added in this commit is
`AbsoluteIndexSet::aocl_range` which returns the range of AOCL leafs
from which the absolute index set *could* have been derived.