security(rpc_server): Change interface type to `TransferTransaction`
This change gives type assurance to callers that they do not leak any
secrets. Specifically, the type `TransferTransaction` guarantees that a
`PrimitiveWitness`, which leaks secret keys, can never be shared with
the RPC server. The type `TransferTransaction` can only ever be a
transaction that is either backed by a `ProofCollection` or a
`SingleProof`.