feat(RPC): Add endpoint `list-utxos`
New endpoint lists all UTXOs, including expected, pending (mempool),
and abandoned UTXOs. It returns this information as a `Vec` of a new
object type `UiUtxo`, which contains:
 - when the UTXO was received, if at all; receipt status otherwise
 - when the UTXO was spent, if at all; spending status otherwise
 - amount
 - release date
 - if available, AOCL leaf index

The motivation for the new type and endpoint is in the dashboard to
compensate for the lost information in the "mempool screen", which
dropped two columns. These columns were useful for getting an idea
of which UTXOs were being used for a transaction. Now the user can
get this information from the UTXOs screen instead.