fix(mutator_set): Undo destruction of AMS' batch_remove
This method was broken in the commits
5acf2073ece8fce4ae41a51a2ff132c1020df7dc and
343209ef9242d42055b01f2588d8e387f474c933 with the horrendous and lazy
rewrites that made the archival mutator set implementation start using
`MutatorSetAccumulator` as a helper library.

Tests have been expanded on and robustified to reveal the damage done in
those two commits.

There are still methods on archival mutator set that goes the lazy way
of constructing a mutator set accumulator and using its method as a
helper method. We should get rid of those and make the archival mutator
set's logic stand-alone, but at least now a bug is removed.