fix(RPC): Filter out duplicate `UiUtxo`s
The RPC call `list-utxos` returns `UiUtxo`s from three sources:
 1. the wallet database of monitored UTXOs;
 2. the wallet database of expected UTXOs;
 3. the mempool.

It is possible for the same `UiUtxo` to be present in more than one of
these sources. Previously that would generate duplicate entries in the
list of `UiUtxo`s returned by `list-utxos`. This commit changes that
behavior by filtering out duplicates.

Furthermore, it changes the order of (2) and (3), so that if the
`UiUtxo` comes from the mempool, it is not marked as "expected" when
happens to also be present in the expected UTXOs database.

Lastly, the signature of mempool function `mempool_unspent_utxos_iter`
is changed to accommodate this duplicate filter.