fork!: Change PoW algorithm to make proposal-switching free
Change the PoW algorithm in order to increase transaction throughput.

A small algorithmical change to the PoW algorithm which constitutes a
hard-fork since old versions of the software will consider new blocks
invalid.

The current PoW algorithm has a preprocessing step that takes a few
seconds on a very fast GPU and ~40 seconds on a fast CPU. This
preprocessing step is, prior to this commit, invalidated by a new block
proposal, meaning that the guessers are heavily disincentivized from
switching block proposals while guessing. The preprocessing step
involves building a large Merkle tree, and it is the leafs of this
Merkle tree that used to depend on the concrete block proposal. With
this commit, the leafs instead depend only on the previous' block's
hash which means that switching between block proposals is free as
long as the parent of the two proposals is the same.

This commit changes only the preprocessing. The tight-loop where the
actual guessing takes place is unchanged for those implementations that
store the entire Merkle tree.

This change also means that preprocessing can start before the block
proof is finished, i.e. before the composer has a block propoosal.
However guessing cannot because the block hash depends on the block
proof.

The reason that this change is good for transaction throughput is that
it makes it much more likely that the competitive miner will switch to
whatever proposal that pays them the highest fee. And merging another
fee-paying transaction into the block proposal increases the reward
for the guesser. In the scenario where a mining pool decides both the
block proof and the guessser reward (and pays out users for their
near-misses) this change *also* benefits transaction throughput since
the preprocessing step (being up to 42GB) will be done locally at each
user (on each GPU or machine), and if they can reuse the preprocessed
data for a new block proposal/task sent by the mining pool, the mining
pool is more likely to update the PoW puzzle regularly since new
transactions can be merged into the coinbase as they appear in the
mempool, without the cost of pausing guessing.