ci: Update `dist` version and settings
With the now-continued development of release tool `dist`, the custom
runners can be removed in order to rely on the (once again) sane
defaults.

In order to reduce the number of surprises during artifact generation,
the new settings tell `dist` to always generate artifacts, not just
come release time.

changelog: ignore