fix(Mempool): Cover edge case for initiating Update
This commit fixes an edge case in the mempool logic that results in transactions
initiated by ourselves to be ignored for Updates as new blocks are found, even
when we were the ones to produce a SingleProof for them.

The edge case comes about as follows:
 - We initiate the transaction and produce a SingleProof for it.
 - We broadcast the transaction.
 - Someone else merges the transaction with some other transaction and
   broadcasts that.
 - That merge-result is inserted into our mempool and this insertion moves the
   original transaction into the cache (without primitive witness).
 - The other operand of the merger is mined. So our transaction is fetched from
   the cache and inserted back into the mempool (without primitive witness).
 - A new block is mined.
 - We do not Update our transaction because
 - it does not have an associated primitive witness <-- error is here.

The correct criterion for deciding whether a transaction should be updated is
whether it is critical; not whether it has a primitive witness. Critical means
"initiated by us". Transactions cannot have a primitive witness without being
critical; but they obviously can be critical without having a primitive witness.
More to the point, the primitive witness is not relevant data for performing an
Update job.

Co-authored-by: Thorkil Schmidiger <thor@neptune.cash>