feat: Refuse connection on bad timestamp
If the peer that the connection is made with reports another timestamp
in the handshake, then the connection is refused.

This change is made because it disallows a DOS'er to reuse the
handshake. More DOS protection should of course be made but this is
already a step in the right direction.

The limit is set to 90 seconds which should be enough on even bad
connections. This limit is in the same order of magnitude as the
standard timeout of a TCP connection, 72 seconds on Windows.

https://serverfault.com/questions/193160/which-is-the-default-tcp-connect-timeout-in-windows