refactor(`GuesserBuffer`): Exfiltrate field `mast_auth_paths`
When switching to a new and improved block proposal, the MAST authentication
paths used for guessing change, but (under the new consensus rule set) the
remainder of the `GuesserBuffer` does not. This commit moves the field
`mast_auth_paths` out of this struct and into the relevant functions' argument
lists. As a result, the `GuesserBuffer` captures exactly that information that
does not change between block proposals and can remain immutable.

Co-authored-by: Thorkil Schmidiger <thor@neptune.cash>